requirements.txt
pandas
prefect
numpy
matplotlib
plotly
seaborn
duckdb
pyopensky
kafka-python
nbformat
scikit-learn
boto3
httpx